+++ src/FileAccessException.php
@@ -17,7 +17,7 @@ namespace Niirrty\IO;
 /**
  * This exception should be used if accessing a file for reading and/or writing fails.
  *
- * The class extends from {@see \Niirrty\IO\IOException}.
+ * The class extends from {@see IOException}.
  *
  * @since v0.1
  */
@@ -30,27 +30,27 @@ class FileAccessException extends IOException
     /**
      * Reading file access.
      */
-    public const ACCESS_READ = 'read';
+    public final const ACCESS_READ = 'read';
 
     /**
      * Writing file access.
      */
-    public const ACCESS_WRITE = 'write';
+    public final const ACCESS_WRITE = 'write';
 
     /**
      * Reading and writing file access.
      */
-    public const ACCESS_READWRITE = 'read and write';
+    public final const ACCESS_READWRITE = 'read and write';
 
     /**
      * Creating file access.
      */
-    public const ACCESS_CREATE = 'create';
+    public final const ACCESS_CREATE = 'create';
 
     /**
      * Deleting file access.
      */
-    public const ACCESS_DELETE = 'delete';
+    public final const ACCESS_DELETE = 'delete';
 
     #endregion
 
@@ -58,16 +58,16 @@ class FileAccessException extends IOException
     #region // = = =   C O N S T R U C T O R   = = = = = = = = = = = = = = = = = = = = = = = = = = = =
 
     /**
-     * Init's a new instance.
+     * Init a new instance.
      *
      * @param string          $file     The file where accessing fails
      * @param string          $access   The access type (see \UK\IO\FileAccessException::ACCESS_* class constants)
      * @param string|null     $message  The optional error message
-     * @param string|int      $code     A optional error code (Defaults to \E_USER_ERROR)
-     * @param \Throwable|null $previous A Optional previous exception.
+     * @param string|int      $code     optional error code (Defaults to \E_USER_ERROR)
+     * @param \Throwable|null $previous Optional previous exception.
      */
     public function __construct(
-        string $file, private string $access = self::ACCESS_READ, ?string $message = null, string|int $code = 256,
+        string      $file, private readonly string $access = self::ACCESS_READ, ?string $message = null, string|int $code = 256,
         ?\Throwable $previous = null )
     {
 
@@ -86,7 +86,7 @@ class FileAccessException extends IOException
     #region // - - -   P U B L I C   M E T H O D S   - - - - - - - - - - - - - - - - - - - - - - - - -
 
     /**
-     * Returns the error file access type (see \UK\IO\FileAccessException::ACCESS_* class costants)
+     * Returns the error file access type (see \UK\IO\FileAccessException::ACCESS_* class constants)
      *
      * @return string
      */
@@ -105,8 +105,8 @@ class FileAccessException extends IOException
      *
      * @param string          $file     The file where reading fails.
      * @param string|null     $message  The optional error message.
-     * @param integer         $code     A optional error code (Defaults to \E_USER_ERROR)
-     * @param \Throwable|null $previous A Optional previous exception.
+     * @param integer         $code     optional error code (Defaults to \E_USER_ERROR)
+     * @param \Throwable|null $previous Optional previous exception.
      *
      * @return FileAccessException
      */
@@ -130,8 +130,8 @@ class FileAccessException extends IOException
      *
      * @param string          $file     The file where reading fails.
      * @param string|null     $message  The optional error message.
-     * @param integer         $code     A optional error code (Defaults to \E_USER_ERROR)
-     * @param \Throwable|null $previous A Optional previous exception.
+     * @param integer         $code     optional error code (Defaults to \E_USER_ERROR)
+     * @param \Throwable|null $previous Optional previous exception.
      *
      * @return FileAccessException
      */