+++ src/FileAlreadyExistsException.php
@@ -15,9 +15,9 @@ namespace Niirrty\IO;
 
 
 /**
- * This class defines a exception, thrown if a file already exists but it should not exist.
+ * This class defines an exception, thrown if a file already exists, but it should not exist.
  *
- * It extends from {@see \Niirrty\IO\IOException}.
+ * It extends from {@see IOException}.
  *
  * @since v0.1
  */
@@ -33,7 +33,7 @@ class FileAlreadyExistsException extends IOException
      * @param string          $file     The already existing file
      * @param string|null     $message  The optional error message
      * @param string|int      $code     The optional error code (Default to \E_USER_ERROR)
-     * @param \Throwable|null $previous A optional previous exception
+     * @param \Throwable|null $previous optional previous exception
      */
     public function __construct(
         string $file, ?string $message = null, string|int $code = \E_USER_ERROR, ?\Throwable $previous = null)