+++ src/FileFormatException.php
@@ -15,9 +15,9 @@ namespace Niirrty\IO;
 
 
 /**
- * This exception should be used if a accessed file contains a wrong/bad format.
+ * This exception should be used if an accessed file contains a wrong/bad format.
  *
- * The class extends from {@see \Niirrty\IO\IOException}.
+ * The class extends from {@see IOException}.
  *
  * @since v0.1
  */
@@ -28,12 +28,12 @@ class FileFormatException extends IOException
     #region // = = =   C O N S T R U C T O R   = = = = = = = = = = = = = = = = = = = = = = = = = = = =
 
     /**
-     * Init's a new instance
+     * Init a new instance
      *
      * @param string          $file     The bad formatted file.
      * @param string|null     $message  The optional error message
      * @param string|int      $code     The optional error code (Default to \E_USER_ERROR)
-     * @param \Throwable|null $previous A optional previous exception
+     * @param \Throwable|null $previous optional previous exception
      */
     public function __construct(
         string $file, ?string $message = null, string|int $code = \E_USER_ERROR, ?\Throwable $previous = null )